refactor: Order/OrderItem Builder 패턴 적용 및 반정규화 보안 강화
- Order, OrderItem에 Builder 패턴 적용
  - @Builder, @NoArgsConstructor, @AllArgsConstructor 추가
  - Order.orderItems에 @Builder.Default로 ArrayList 초기화 보장

- OrderService.processOrderCreation() 리팩토링
  - Setter 방식에서 Builder 패턴으로 변경
  - 보안 강화: unitPrice, totalPrice를 프론트엔드가 아닌 백엔드에서 계산
  - Item.getPrice()에서 자동으로 unitPrice 복사 (반정규화)
  - 가격 조작 방지 및 데이터 정합성 보장

- OrderRequestDto.OrderItemDto 간소화
  - itemName, unitPrice, totalPrice, itemCategory 제거
  - itemId, quantity, serialNum만 전달받도록 변경
  - 프론트엔드 책임 최소화

- Delivery 엔티티 정리
  - recipientName, recipientPhone, recipientAddr 컬럼 주석 처리
  - Customer 정보로 대체 (중복 제거)

- Item 엔티티 메서드명 정리
  - getSellingPrice() 관련 정리

+++ BackEnd/pcgear/src/main/java/com/pcgear/complink/pcgear/Item/Item.java
@@ -1,5 +1,7 @@
 package com.pcgear.complink.pcgear.Item;
 
+import java.math.BigDecimal;
+
 import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.persistence.Convert;
 import jakarta.persistence.Entity;
@@ -39,9 +41,9 @@ public class Item {
 
         @NotNull(message = "입고단가는 필수 입력 항목입니다.")
         @PositiveOrZero(message = "입고단가는 0 또는 양수여야 합니다.")
-        private Integer purchasePrice;
+        private BigDecimal purchasePrice;
 
         @NotNull(message = "출고단가는 필수 입력 항목입니다.")
         @PositiveOrZero(message = "출고단가는 0 또는 양수여야 합니다.")
-        private Integer sellingPrice;
+        private BigDecimal sellingPrice;
 }
